refactor the two existing switch statements into polymorphic dispatch
switch statement in condigureCell is replaced with AnyCellConfigurable.configure which force-casts and calls the type-specific CellConfigurable.configure

we can assume all cells are of type AnyCellConfigurable because we register only using the cellTypes array, and that only allows AnyCellConfigurable.Type.

the switch statement in cellType(for:) is replaced by the fact that CellViewModelType has an associatedtype of the cell type is can configure.